+++ Content/Items/Food/Seasoning.TableSalt.cs
@@ -9,6 +9,7 @@ namespace StarlightRiver.Content.Items.Food
 		public override void SafeSetDefaults()
 		{
 			Item.createTile = ModContent.TileType<Tiles.Cooking.TableSalt>();
+			Item.consumable = true;
 			Item.value = 400;
 			Item.rare = ItemRarityID.White;
 		}
