+++ Content/Tiles/Cooking/Salts.cs
@@ -34,8 +34,6 @@ namespace StarlightRiver.Content.Tiles.Cooking
 			TileObjectData.addTile(Type);
 		}
 
-		//It's stupid, but this is nessecary to make nonsolid tiles able to be placed next to other nonsolid tiles
-		#region tilehacking
 		public override bool CanPlace(int i, int j)
 		{
 			Tile tile = Framing.GetTileSafely(i, j + 1);
@@ -74,7 +72,6 @@ namespace StarlightRiver.Content.Tiles.Cooking
 			Tile tile = Framing.GetTileSafely(i, j);
 			Main.tileSolid[tile.TileType] = false;
 		}
-		#endregion
 
 		public override void NearbyEffects(int i, int j, bool closer)
 		{
@@ -144,131 +141,6 @@ namespace StarlightRiver.Content.Tiles.Cooking
 		}
 	}
 
-	internal class TableSaltProjectile : ModProjectile
-	{
-		protected bool falling = true;
-
-		protected virtual int TileType => ModContent.TileType<TableSalt>();
-
-		public virtual int DustType => DustID.Marble;
-
-		public override string Texture => AssetDirectory.CookingTile + Name;
-
-		public override void SetStaticDefaults()
-		{
-			DisplayName.SetDefault("Sea Salt");
-			ProjectileID.Sets.ForcePlateDetection[Projectile.type] = true;
-		}
-
-		public override void SetDefaults()
-		{
-			Projectile.knockBack = 6f;
-			Projectile.width = 10;
-			Projectile.height = 10;
-			Projectile.friendly = true;
-			Projectile.hostile = true;
-			Projectile.penetrate = -1;
-		}
-
-		public override void AI()
-		{
-			if (Main.rand.NextBool(5))
-			{
-				int dust = Dust.NewDust(Projectile.position, Projectile.width, Projectile.height, DustType);
-				Main.dust[dust].velocity.X *= 0.4f;
-			}
-
-			Projectile.tileCollide = true;
-			Projectile.localAI[1] = 0f;
-
-			if (Projectile.ai[0] == 1f)
-			{
-				if (!falling)
-				{
-					Projectile.ai[1] += 1f;
-
-					if (Projectile.ai[1] >= 60f)
-					{
-						Projectile.ai[1] = 60f;
-						Projectile.velocity.Y += 0.2f;
-					}
-				}
-				else
-				{
-					Projectile.velocity.Y += 0.41f;
-				}
-			}
-			else if (Projectile.ai[0] == 2f)
-			{
-				Projectile.velocity.Y += 0.2f;
-
-				if (Projectile.velocity.X < -0.04f)
-					Projectile.velocity.X += 0.04f;
-				else if (Projectile.velocity.X > 0.04f)
-					Projectile.velocity.X -= 0.04f;
-				else
-					Projectile.velocity.X = 0f;
-			}
-
-			Projectile.rotation += 0.1f;
-
-			if (Projectile.velocity.Y > 10f)
-				Projectile.velocity.Y = 10f;
-		}
-
-		public override bool TileCollideStyle(ref int width, ref int height, ref bool fallThrough, ref Vector2 hitboxCenterFrac)
-		{
-			if (falling)
-				Projectile.velocity = Collision.AnyCollision(Projectile.position, Projectile.velocity, Projectile.width, Projectile.height, true);
-			else
-				Projectile.velocity = Collision.TileCollision(Projectile.position, Projectile.velocity, Projectile.width, Projectile.height, fallThrough, fallThrough, 1);
-
-			return false;
-		}
-
-		public override void Kill(int timeLeft)
-		{
-			if (Projectile.owner == Main.myPlayer && !Projectile.noDropItem)
-			{
-				int tileX = (int)(Projectile.position.X + Projectile.width / 2) / 16;
-				int tileY = (int)(Projectile.position.Y + Projectile.width / 2) / 16;
-
-				Tile tile = Main.tile[tileX, tileY];
-				Tile tileBelow = Main.tile[tileX, tileY + 1];
-
-				if (tile.BlockType == BlockType.HalfBlock && Projectile.velocity.Y > 0f && System.Math.Abs(Projectile.velocity.Y) > System.Math.Abs(Projectile.velocity.X))
-					tileY--;
-
-				if (!tile.HasTile)
-				{
-					bool onMinecartTrack = tileY < Main.maxTilesY - 2 && tileBelow != null && tileBelow.HasTile && tileBelow.TileType == TileID.MinecartTrack;
-
-					if (!onMinecartTrack)
-						WorldGen.PlaceTile(tileX, tileY, TileType, false, true);
-
-					if (!onMinecartTrack && tile.HasTile && tile.TileType == TileType)
-					{
-						if (tileBelow.BlockType == BlockType.HalfBlock || tileBelow.Slope != 0)
-						{
-							WorldGen.SlopeTile(tileX, tileY + 1, 0);
-
-							if (Main.netMode == NetmodeID.Server)
-								NetMessage.SendData(MessageID.TileManipulation, -1, -1, null, 14, tileX, tileY + 1);
-						}
-
-						if (Main.netMode != NetmodeID.SinglePlayer)
-							NetMessage.SendData(MessageID.TileManipulation, -1, -1, null, 1, tileX, tileY, TileType);
-					}
-				}
-			}
-		}
-
-		public override bool? CanDamage()
-		{
-			return Projectile.localAI[1] != -1f;
-		}
-	}
-
 	internal class PinkSeaSalt : TableSalt
 	{
 		public override int DustType => DustID.Orichalcum;
@@ -280,10 +152,25 @@ namespace StarlightRiver.Content.Tiles.Cooking
 		public override int ItemDrop => ModContent.ItemType<Items.Food.SeaSalt>();
 	}
 
-	internal class PinkSeaSaltProjectile : TableSaltProjectile
+	internal class PinkSeaSaltProjectile : FallingTileProjectile
 	{
 		protected override int TileType => ModContent.TileType<PinkSeaSalt>();
 
-		public override int DustType => DustID.Orichalcum;
+		protected override int DustType => DustID.Orichalcum;
+
+		protected override string ProjectileName => "Sea Salt";
+
+		public override string Texture => AssetDirectory.CookingTile + Name;
+	}
+
+	internal class TableSaltProjectile : FallingTileProjectile
+	{
+		protected override int TileType => ModContent.TileType<TableSalt>();
+
+		protected override int DustType => DustID.Marble;
+
+		protected override string ProjectileName => "Table Salt";
+
+		public override string Texture => AssetDirectory.CookingTile + Name;
 	}
 }
